refactor(cron): share managed block helpers across setup scripts
